third_party/WebKit/LayoutTests/animations/events/animation-events-create.html

<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>

description("Test dynamic creation of AnimationEvent.");

if (document.createEvent)
    testPassed("document.createEvent exists");
else
    testFailed("document.createEvent missing");

debug("");
debug("Test creation of AnimationEvent");

var ev = document.createEvent("AnimationEvent");

shouldBe("typeof(ev)", "'object'");
shouldBe("ev.animationName", "''");
shouldBe("ev.elapsedTime", "0.0");

debug("");
debug("Create a MouseEvent and make sure it doesn't have AnimationEvent properties");

ev = document.createEvent("MouseEvent");

shouldBe("typeof(ev)", "'object'");
shouldBe("ev.animationName", "undefined");
shouldBe("ev.elapsedTime", "undefined");

debug("");

</script>
</body>
</html>
