third_party/WebKit/LayoutTests/animations/events/animation-events-create-expected.txt

Test dynamic creation of AnimationEvent.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.createEvent exists


Test creation of AnimationEvent
PASS typeof(ev) is 'object'
PASS ev.animationName is ''
PASS ev.elapsedTime is 0.0


Create a MouseEvent and make sure it doesn't have AnimationEvent properties
PASS typeof(ev) is 'object'
PASS ev.animationName is undefined
PASS ev.elapsedTime is undefined


PASS successfullyParsed is true

TEST COMPLETE
